pyproject.toml
[tool.poetry]
name = "web-chat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
unicorn = "^2.0.1.post1"
fastapi = {extras = ["all"], version = "^0.104.1"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
pydantic = "^2.5.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
